# artifacts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "共域 CoField API",
    "description": "以共同事件为核心的校园社交连接器。\n\n领域词汇见 CONTEXT.md；用户可见文案不使用这些词汇，映射见 docs/07。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/me/profile": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Read",
        "description": "我这边现在是什么样。\n\n身份第一次出现时 `get_repositories` 已经把行建好了，所以这里不会\n遇到\"人不存在\"——那一步放在依赖里而不是每个端点各自记得做。",
        "operationId": "read_api_me_profile_get",
        "parameters": [
          {
            "name": "x-principal-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Principal-Id"
            }
          },
          {
            "name": "x-campus-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Campus-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "Write",
        "description": "改写我这边。整份覆盖，不是追加——\"我不再想参与拍摄了\"必须说得出口。",
        "operationId": "write_api_me_profile_put",
        "parameters": [
          {
            "name": "x-principal-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Principal-Id"
            }
          },
          {
            "name": "x-campus-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Campus-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProfileIn": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          },
          "open_to": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Open To"
          },
          "self_intro": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Self Intro"
          },
          "zone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone"
          },
          "major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Major"
          }
        },
        "type": "object",
        "title": "ProfileIn"
      },
      "ProfileOut": {
        "properties": {
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "named_self": {
            "type": "boolean",
            "title": "Named Self",
            "default": true
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          },
          "open_to": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Open To"
          },
          "self_intro": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Self Intro"
          },
          "zone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone"
          },
          "major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Major"
          },
          "not_recognised": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Not Recognised"
          }
        },
        "type": "object",
        "required": [
          "display_name",
          "skills",
          "open_to"
        ],
        "title": "ProfileOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}